+++ silkworm/capi/instance.hpp
@@ -22,7 +22,7 @@
 
 #include <boost/asio/cancellation_signal.hpp>
 
-#include <silkworm/db/snapshots/repository.hpp>
+#include <silkworm/db/snapshots/snapshot_repository.hpp>
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/infra/concurrency/context_pool_settings.hpp>
 #include <silkworm/node/stagedsync/execution_engine.hpp>
