+++ silkworm/db/blocks/headers/header_index.hpp
@@ -22,7 +22,7 @@
 
 #include <silkworm/core/common/bytes.hpp>
 #include <silkworm/db/snapshots/index_builder.hpp>
-#include <silkworm/db/snapshots/path.hpp>
+#include <silkworm/db/snapshots/snapshot_path.hpp>
 #include <silkworm/infra/common/memory_mapped_file.hpp>
 
 namespace silkworm::snapshots {
