+++ silkworm/db/snapshot_merger.cpp
@@ -24,9 +24,9 @@
 #include <silkworm/infra/common/filesystem.hpp>
 #include <silkworm/infra/common/log.hpp>
 
-#include "snapshots/path.hpp"
 #include "snapshots/seg/compressor.hpp"
 #include "snapshots/snapshot_bundle.hpp"
+#include "snapshots/snapshot_path.hpp"
 #include "snapshots/snapshot_writer.hpp"
 
 namespace silkworm::db {
