+++ silkworm/db/snapshots/snapshot_bundle_factory.hpp
@@ -24,8 +24,8 @@
 #include <silkworm/core/common/base.hpp>
 
 #include "index_builder.hpp"
-#include "path.hpp"
 #include "snapshot_bundle.hpp"
+#include "snapshot_path.hpp"
 #include "snapshot_type.hpp"
 
 namespace silkworm::snapshots {
