+++ silkworm/db/test_util/temp_snapshots.hpp
@@ -21,8 +21,8 @@
 
 #include <silkworm/core/common/endian.hpp>
 #include <silkworm/core/common/util.hpp>
-#include <silkworm/db/snapshots/repository.hpp>
 #include <silkworm/db/snapshots/seg/common/varint.hpp>
+#include <silkworm/db/snapshots/snapshot_repository.hpp>
 #include <silkworm/infra/test_util/temporary_file.hpp>
 
 namespace silkworm::snapshots::test_util {
