+++ silkworm/db/transactions/txn_index.hpp
@@ -23,7 +23,7 @@
 #include <silkworm/core/common/bytes.hpp>
 #include <silkworm/db/etl/collector.hpp>
 #include <silkworm/db/snapshots/index_builder.hpp>
-#include <silkworm/db/snapshots/path.hpp>
+#include <silkworm/db/snapshots/snapshot_path.hpp>
 #include <silkworm/infra/common/memory_mapped_file.hpp>
 
 namespace silkworm::snapshots {
